requirements.txt
asgiref==3.5.2
certifi==2022.9.24
cffi==1.15.1
chardet==3.0.4
charset-normalizer==2.1.1
colorama==0.4.5
cryptography==38.0.1
Django==4.0
django-cors-headers==3.13.0
django-rest-knox==4.2.0
djangorestframework==3.14.0
et-xmlfile==1.1.0
googletrans==3.0.0
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2022.10.1
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
numpy==1.23.3
openai==0.23.1
openpyxl==3.0.10
pandas==1.5.0
pandas-stubs==1.5.0.221003
pycparser==2.21
python-dateutil==2.8.2
pytz==2022.4
requests==2.28.1
rfc3986==1.5.0
six==1.16.0
sniffio==1.3.0
sqlparse==0.4.3
tqdm==4.64.1
gunicorn==20.1.0
types-pytz==2022.4.0.0
typing_extensions==4.3.0
tzdata==2022.4
urllib3==1.26.12
dj-database-url==1.0.0
psycopg2-binary==2.9.4
whitenoise==6.2.0
Brotli==1.0.9
django-rest-swagger==2.2.0
pyyaml 
uritemplate